+++ src/backend/interfaces/REST/py.server/caliopen_api/user/config.py
@@ -23,3 +23,6 @@ def includeme(config):
 
     log.debug('Loading device API')
     config.scan('caliopen_api.user.device')
+
+    log.debug('Loading imports API')
+    config.scan('caliopen_api.user.imports')
